chatbot-model/requirements.txt
sentence-transformers>=2.3.0
transformers>=4.35.0
torch>=2.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0
huggingface-hub>=0.20.0
datasets>=2.14.0
accelerate>=0.26.0
rank-bm25>=0.2.2
groq>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
